backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai
google-genai
python-dotenv
aiofiles
pypdf
webauthn
google-auth
requests
PyJWT
pywebpush
apscheduler
slowapi
cryptography
httpx
